requirements.txt
beautifulsoup4==4.12.3
chromadb==0.5.0
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
python-dotenv==1.0.1
Requests==2.31.0
scikit_learn==1.4.2
SQLAlchemy==2.0.29
streamlit==1.34.0
